Export as a MD file is added.

+++ src/Logges/utils.py
@@ -7,8 +7,8 @@
 """
 import datetime
 import os
-import platform
 import sys
+import platform
 from io import TextIOWrapper
 from typing import Dict
 from typing import List
@@ -86,7 +86,8 @@ def create_pie_chart(saving_path: str, status_dict: Dict[str, int]) -> None:
 
     png_path = os.path.join(saving_path, "pie_chart.png")
 
-    os.remove(png_path)
+    if os.path.exists("pie_chart.png"):
+        os.remove(png_path)
 
     plt.savefig(f"{png_path}")
 
@@ -202,6 +203,104 @@ def console_data(script_name: str, status_dict: Dict[str, int],
         f"\tCRITICAL: %{round(status_dict['CRITICAL']/total_length*100 if total_length > 0 else 0, 2)}")
 
 
+def to_markdown(
+    script_name: str,
+    saving_path: str,
+    status_dict: Dict[str, int],
+    status_icons: Dict[str, str],
+    local_file: bool = False,
+) -> None:
+    """Export the logs to a file with `.pdf` format.
+
+    Parameters:
+        script_name `str`: Save the pdf file as that string.
+        saving_path `str`:  Save the pdf file to the that path.
+        stats_dict `Dict`:  Define status counter.
+        local_file: `bool`: Check load from root directory or local file. Default is False.
+
+    Return:
+        None
+    """
+    if local_file:
+        md_file = os.path.join(
+            saving_path,
+            script_name.replace('.log', '.md')
+        )
+
+        markdown_file = open(md_file, "w")
+        filename = script_name
+
+    else:
+        md_file = os.path.join(
+            saving_path,
+            get_daily_log_file_name(filename=script_name, markdown=True)
+        )
+
+        markdown_file = open(md_file, "w")
+        filename = get_daily_log_file_name(filename=script_name)
+
+    file_dir = saving_path
+    full_logfile_path = os.path.join(file_dir, filename)
+
+    only_filename = "_".join(filename.split("_")[1:]) + ".py".replace(
+        ".log", "")
+    file_date = filename.split("_")[0]
+
+    # Fix Windows Problems.
+    if get_current_platform_name() == "Windows":
+        only_filename = os.path.split(only_filename)[1]
+
+    markdown_file.writelines(
+        f"# {only_filename} {file_date} Logs :see_no_evil: :hear_no_evil: :speak_no_evil:\n"
+    )
+    markdown_file.writelines("![](pie_chart.png)\n")
+    markdown_file.writelines(
+        "|TIME|STATUS|FILENAME|FUNCTION|MESSAGE|\n| :--: | :--: | :--: | :--: | :--: |\n"
+    )
+
+    # Split Strings.
+    file = open(full_logfile_path, "r")
+    (
+        _date_list,
+        _status_list,
+        _filename_list,
+        _functname_list,
+        _log_message_list,
+    ) = extract_logs(logs=file)
+
+    # Write logs in markdown file.
+    for index, _ in enumerate(_log_message_list):
+        log_status_clear = _status_list[index].replace("[", "").replace(
+            "]", "")
+        status_dict[log_status_clear] += 1
+
+        try:
+            markdown_file.writelines("|{}|{}|{}|{}|{}|".format(
+                _date_list[index],
+                status_icons[log_status_clear],
+                _filename_list[index],
+                _functname_list[index],
+                _log_message_list[index].replace("\n", " "),
+            ))
+            markdown_file.write("\n")
+        except KeyError:
+            raise ("Please check your icon.")
+
+    # Write signature
+    markdown_file.writelines(
+        "All right reserved 2022 &copy;&nbsp; [Logges](https://github.com/uysalserkan/Logges) - \
+*[uysalserkan](https://github.com/uysalserkan/) & [Ozkan](https://github.com/ozkanuysal)*\n"
+    )
+
+    # Create chart
+    create_pie_chart(
+        saving_path=saving_path,
+        status_dict=status_dict,
+    )
+
+    markdown_file.close()
+
+
 def to_pdf(
     script_name: str,
     saving_path: str,
@@ -214,6 +313,7 @@ def to_pdf(
         script_name `str`: Save the pdf file as that string.
         saving_path `str`:  Save the pdf file to the that path.
         stats_dict `Dict`:  Define status counter.
+        local_file: `bool`: Check load from root directory or local file. Default is False.
 
     Return:
         None
